purchase_master_service/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.14.0
requests==2.32.3
djangorestframework-simplejwt==5.3.1